# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precision Harvest - Smart Irrigation API
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklyUsageItem } from './weeklyUsageItem';

export interface Analytics {
  farmId: string;
  totalWaterSavedLiters: number;
  moneySaved: number;
  cropHealthIndex: number;
  irrigationEvents: number;
  skippedEvents: number;
  weeklyUsage: WeeklyUsageItem[];
}