# lib/api-zod/src/generated/types/automationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precision Harvest - Smart Irrigation API
 * OpenAPI spec version: 0.1.0
 */

export interface AutomationResult {
  farmId: string;
  action: string;
  reason: string;
  waterSavedLiters: number;
  notificationSent: boolean;
}