# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precision Harvest - Smart Irrigation API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  farmLocation?: string;
}