# lib/api-zod/src/generated/types/weeklyUsageItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Precision Harvest - Smart Irrigation API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyUsageItem {
  day: string;
  traditional: number;
  optimized: number;
  rainfall: number;
}